requirements.txt
anyio==4.9.0
APScheduler==3.11.0
cachetools==4.2.2
certifi==2025.4.26
charset-normalizer==3.4.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
joblib==1.5.1
loguru==0.7.3
numpy==2.3.0
python-telegram-bot==22.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
scikit-learn==1.7.0
scipy==1.15.3
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
threadpoolctl==3.6.0
tornado==6.1
typing_extensions==4.14.0
tzlocal==5.3.1
urllib3==2.4.0
